requirements.txt
# Core Dependencies
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Pillow==10.0.1

# AI/ML Libraries
openai==1.3.7
google-generativeai==0.3.2
protobuf==3.20.3

# Web Framework
streamlit==1.28.1
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5
streamlit-aggrid==0.3.4.post3

# Visualization
plotly==5.17.0

# Export and Reporting
reportlab==4.0.7
fpdf2==2.7.6

# Configuration
python-dotenv==1.0.0

# Optional: Text-to-Speech
pyttsx3==2.90

# Additional dependencies for Google Generative AI
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
